pyproject.toml
[tool.poetry]
name = "navi-parser"
version = "0.1.0"
description = ""
authors = ["suk-6 <me@suk.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^24.10.0"
tqdm = "^4.67.1"
setuptools = "^75.6.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
opencv-python = "^4.10.0.84"
boto3 = "^1.35.74"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
